Implemented docs versioning
The first versioned docs are targeted at Rancher Desktop v1.6.1.

The versioned docs directories for both EN and ZH have been generated and are the default ones when browsing the docs site.

The "Next" version, contains the docs for the future release and once a new version is created, the `/docs` directory is copied to the versioned_docs with the new version number.

Signed-off-by: Nuno do Carmo <nuno.carmo@suse.com>