Added images to /docs and updated all links
Moved the directory `img` under `/docs`, so it won't be affect by images of future releases.

Both the images from the main UI and the preferences have been added.

For readability, the images have been split into `tabs`, so only one is displayed at the time.
The `tabs` are synchronized with the "os tabs group".

Signed-off-by: Nuno do Carmo <nuno.carmo@suse.com>