Document auto-start and background behavior settings
The old "Behavior" tab is now called "General", and the new options
are configurable on the second tab called "Behavior".

The documentation for the new Behavior tab has been duplicated from
Windows to macOS and Linux. The macOS verbiage has been slightly
adjusted to use OS-specific terms, like "app switcher" and "dock"
instead of "task switcher" and "task bar". It also adds "in the
menu bar" to "notification icon" because the concept is normally known
as the "menu bar icon" on macOS.

The Linux version still uses Windows terminology; not sure if there
is native terminology that is consistent across window managers.

This commit also moves the screenshots to the top of the section
describing the individual settings.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>