Move screenshots before their descriptions
In manuals you want to see the illustration first, and then read the
description of the elements in the image. The general structure should be:

* introduction/overview
* screenshot
* detailed description of each element in the screenshot

This commit only moves text around, but doesn't change any.

It also normalizes some of the headings (for Preferences only):

Tab names are level 3 headings
Settings labels are level 4 headings

Signed-off-by: Jan Dubois <jan.dubois@suse.com>