Refine Dockerfile to copy only index.html
Update the Dockerfile to copy only the index.html file instead of the entire directory. 
While the performance impact is minimal for small files, this change aligns with 
Docker best practices and prepares for potential future project growth. It also 
serves as a reminder to be mindful of what we include in our Docker images.

Signed-off-by: Peter Matogo <pmatogo@gmail.com>