Automate versioned image URLs
This implements a remark plugin that scans for all images using URLs
starting with `rd-versioned-asset://` and replace it with a versioned S3
URL in the form of `https://suse-rancher-media.s3.amazonaws.com/desktop/v1`
so that we don't need to manually change the version every release (and
potentially lead to errors).

For `latest` and `next`, we use the largest version listed in
`versions.json`.

The generated files should look identical to before the change.

Signed-off-by: Mark Yen <mark.yen@suse.com>