Update faq.md
with the current latest version that I installed on my mac, I don't find `/var/run/docker.sock` but `~/.rd/docker.sock`.

therefore this document seems wrong but I can't not be sure enough to provide a revised version of the document.

I'm just using this PR to flag the issue and hopefully someone who can update this to update it.

If you would like to reuse this PR to merge, feel free to suggest the change with Github PR's feature. Otherwise, feel free to create a PR to replace this PR please.

Signed-off-by: Heechul Ryu <ryuhcii@gmail.com>